feat: add realtime workers and exports
